Progress meter format do not contain \r anymore
The implementation details of the progress meter should not be the
responsibility of the end user. Therefore, the end user should not have
to deal with the \r at the beginning of its custom format for the
progress meter. This is especially True as having the \r at the
beginning or at the end of the format may cause different behavior.

This commit removes the responsibility of the \r from the end user.
Instead, it adds a `dynamic` keyword to the ProgressMeter class. If set
to True (default), the \r is added at the beginning. Else, the format is
left untouched.