Fix mystery readings in read_tpxheader
read_tpxheader reads the header of a gromacs TPR file. This function
follows the C function do_tpxheader in fileio/tpxio.cpp.

Two integer were read in the python version without apparent
correspondence in the original C function from gromacs. It appears that
these integer are in fact the length of the strings read right after.
The reading that integer is hidden deep in the gromacs source code in
gmx_system_xdr.c:454 (gromacs-2016).

This shows a discrepancy in the way python and gromacs read strings from
a XDR file. The new python function do_string emulate gromacs's
gmx_fio_do_string, and read the size prior to the string.